Disable RADIO when SD margin is reached.
It is possible that the core does not disable RADIO on the margin event
due to critical section. If RADIO event is pending in this situation,
it is not cleared at all and the driver hangs in infinite RADIO event
handler.
This commit disables the RADIO peripheral when SD margin is reached to
prevent RADIO event handling.